Adding ability to get port from telemetry (default 23 for boards). If port is not 23, assume a mock board and post commands instead of opening sockets.

+++ main.js
@@ -15,6 +15,8 @@ var app = express();
 var net = require('net');
 
 var http = require('http').Server(app);
+var httpbase = require('http');
+
 var io = require('socket.io')(http);
 
 var bodyParser = require('body-parser');
@@ -26,26 +28,50 @@ var writeToClient = function(board_id, message) {
 	console.log(message);
 	var clientIP = '';
 	clientIP = telemetry[board_id].ip;
+	clientPort = telemetry[board_id].port;
 	console.log(clientIP);
+	console.log(clientPort);
 	if (clientIP === '') {
 	}
 	else {
-		var client = new net.Socket();
-		client.connect(23, clientIP, function() {
-			console.log('connected to ' + board_id);
-			client.write(message);
-		});
-	
-		client.on('data', function(data) {
-			console.log('data: ' + data);
-			client.destroy();
-		});
-		client.on('close', function() {
-			console.log('connection closed');
-		});
-		client.on('error', function(err) {
-			console.log('error: ' + err.message);
-		});
+		if (clientPort !== 23) {
+			var post_options = {
+				host: clientIP,
+				port: clientPort,
+				path: '/' + message,
+				method: 'POST',
+				headers: {
+					'Content-Type': 'application/x-www-form-urlencoded',
+					'Content-Length': 0
+				}
+			};
+			var post_req = httpbase.request(post_options, function(res) {
+				res.setEncoding('utf8');
+				res.on('data', function(chunk) {
+					console.log('Response: ' + chunk);
+				});
+			});
+			post_req.write("");
+			post_req.end();
+		}
+		else {
+			var client = new net.Socket();
+			client.connect(clientPort, clientIP, function() {
+				console.log('connected to ' + board_id);
+				client.write(message);
+			});
+		
+			client.on('data', function(data) {
+				console.log('data: ' + data);
+				client.destroy();
+			});
+			client.on('close', function() {
+				console.log('connection closed');
+			});
+			client.on('error', function(err) {
+				console.log('error: ' + err.message);
+			});
+		}
 	}
 };
 
@@ -183,6 +209,12 @@ app.post('/status', function(req, res) {
 		timeouts[sname] = 0;
 		telemetry[sname].connection = "active";
 		telemetry[sname].ip = req.ip;
+		if ('port' in req.headers) {
+			telemetry[sname].port = req.headers.port;
+		}
+		else {
+			telemetry[sname].port = 23;
+		}
 		telemetry[sname].firmver = req.headers.fver;
 		telemetry[sname].swarm = req.headers.sw_arm;
 		telemetry[sname].hwarm = req.headers.hw_arm;
